# API-SPEC.yaml
openapi: 3.0.3
info:
  title: SFOps API
  description: Salesforce Deployment SaaS Platform API
  version: 1.0.0
  contact:
    name: SFOps Support
    email: support@sfops.io
servers:
  - url: https://api.sfops.io/v1
    description: Production
  - url: https://staging-api.sfops.io/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - bearerAuth: []
...
paths:
...
  /tenants/{tenantId}/orgs:
    get:
      tags: [Orgs]
      summary: List connected Salesforce orgs
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: List of orgs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Org'

    post:
      tags: [Orgs]
      summary: Create org connection (returns OAuth URL)
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                environment:
                  type: string
                  enum: [production, sandbox, scratch]
                instanceUrl:
                  type: string
      responses:
        '201':
          description: OAuth initiation URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgId:
                    type: string
                  oauthUrl:
                    type: string
                  state:
                    type: string
  /tenants/{tenantId}/orgs/{orgId}:
    get:
      tags: [Orgs]
      summary: Get org details
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/orgId'
      responses:
        '200':
          description: Org details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'

    delete:
      tags: [Orgs]
      summary: Disconnect org
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/orgId'
      responses:
        '204':
          description: Org disconnected
  /tenants/{tenantId}/orgs/{orgId}/oauth/callback:
    post:
      tags: [Orgs]
      summary: OAuth callback for org connection
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/orgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: Org connected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'

  # Snapshots
...
  /tenants/{tenantId}/diffs:
    post:
      tags: [Diffs]
      summary: Create metadata comparison
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceOrgId, targetOrgId]
              properties:
                sourceOrgId:
                  type: string
                targetOrgId:
                  type: string
                sourceSnapshotId:
                  type: string
                targetSnapshotId:
                  type: string
      responses:
        '200':
          description: Diff results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Diff'
  /tenants/{tenantId}/diffs/{diffId}:
    get:
      tags: [Diffs]
      summary: Get diff details
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/diffId'
      responses:
        '200':
          description: Diff details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Diff'

  # Packages
...
  /tenants/{tenantId}/deploys:
    get:
      tags: [Deployments]
      summary: List deployments
      parameters:
        - $ref: '#/components/parameters/tenantId'
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'

    post:
      tags: [Deployments]
      summary: Start deployment
      parameters:
        - $ref: '#/components/parameters/tenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [packageId, targetOrgId]
              properties:
                packageId:
                  type: string
                targetOrgId:
                  type: string
                runTests:
                  type: boolean
                checkOnly:
                  type: boolean
                  description: Validation only (dry-run)
                testLevel:
                  type: string
                  enum: [NoTestRun, RunSpecifiedTests, RunLocalTests, RunAllTestsInOrg]
      responses:
        '202':
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
  /tenants/{tenantId}/deploys/{deployId}:
    get:
      tags: [Deployments]
      summary: Get deployment status
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/deployId'
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
  /tenants/{tenantId}/deploys/{deployId}/status:
    get:
      tags: [Deployments]
      summary: Get real-time deployment status (SSE)
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/deployId'
      responses:
        '200':
          description: Server-sent events stream
          content:
            text/event-stream:
              schema:
                type: object
  /tenants/{tenantId}/deploys/{deployId}/rollback:
    post:
      tags: [Deployments]
      summary: Rollback deployment
      parameters:
        - $ref: '#/components/parameters/tenantId'
        - $ref: '#/components/parameters/deployId'
      responses:
        '202':
          description: Rollback initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'

  # Backups
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
...
    diffId:
      name: diffId
      in: path
      required: true
      schema:
        type: string
...
    deployId:
      name: deployId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Org:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        environment:
          type: string
          enum: [production, sandbox, scratch]
        instanceUrl:
          type: string
        username:
          type: string
        orgId:
          type: string
        status:
          type: string
          enum: [connected, disconnected, expired]
        lastSyncedAt:
          type: string
          format: date-time
...
    Diff:
      type: object
      properties:
        id:
          type: string
        sourceOrgId:
          type: string
        targetOrgId:
          type: string
        changes:
          type: array
          items:
            $ref: '#/components/schemas/MetadataChange'
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/Dependency'
    MetadataChange:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        changeType:
          type: string
          enum: [added, modified, deleted]
        sourceHash:
          type: string
        targetHash:
          type: string
    Dependency:
      type: object
      properties:
        component:
          type: string
        dependsOn:
          type: array
          items:
            type: string
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        packageId:
          type: string
        targetOrgId:
          type: string
        status:
          type: string
          enum: [pending, validating, running, succeeded, failed, rolled_back]
        checkOnly:
          type: boolean
        progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeploymentResult'
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    DeploymentResult:
      type: object
      properties:
        componentType:
          type: string
        componentName:
          type: string
        status:
          type: string
          enum: [pending, succeeded, failed, skipped]
        message:
          type: string
...